+++ lib/Utils/index.d.ts
@@ -53,5 +53,6 @@ export * from "./history-share.js";
 export * from "./transcribe.js";
 export * from "./auth-secure.js";
 export * from "./qr-render.js";
+export * from "./auto-reconnect.js";
 
 export * from "./MessageBuilder.js";